Switch out russian puzzles for adjusted ones
Use a combo of both the dictionary and also frequency sorted lists.

Reasoning:

Frequency sorted lists include emojis and non-words, which we don't
want.

The dictionary is quite large when you consider words in all their
conjugations (it's about a million lines). A lot of these conjugations
aren't commonly used.

So instead go through the top 500k used russian words on the internet
and filter out any non-words using the dictionary as a reference.